refactor: sort code and partial split of decorator logic
WIP
