+++ backend/funix/decorator/file.py
@@ -3,19 +3,43 @@ HTTP File service for funix.
 """
 
 from io import BytesIO
-from os.path import abspath, splitext
+from os.path import abspath, join, splitext
+from threading import Timer
 from typing import Any
 from uuid import uuid4
 
 from flask import abort, send_file
 
+from funix import create_safe_tempdir
 from funix.app import app
+from funix.config.switch import GlobalSwitchOption
 from funix.util.uri import is_valid_uri
 
 __files_dict: dict[str, bytes | str] = {}
 """
 A dict, key is file id, value is file content (path or bytes).
 """
+
+
+def delete_file(fid: str) -> None:
+    """
+    Delete the file.
+
+    Parameters:
+        fid (str): The file id.
+
+    Note:
+        Need lock, but it's not a big problem for now I think so?
+    """
+    global __files_dict
+    if fid in __files_dict:
+        del __files_dict[fid]
+
+
+def delete_file_task(fid: str) -> None:
+    if GlobalSwitchOption.FILE_LINK_EXPIRE_TIME != -1:
+        t = Timer(GlobalSwitchOption.FILE_LINK_EXPIRE_TIME, delete_file, args=(fid,))
+        t.start()
 
 
 def get_real_uri(path_or_file_content: str | bytes) -> str:
@@ -36,10 +60,20 @@ def get_real_uri(path_or_file_content: str | bytes) -> str:
     if isinstance(path_or_file_content, bytes):
         fid = uuid4().hex
         result = f"/file/{fid}"
-        if path_or_file_content not in list(__files_dict.values()):
-            __files_dict[fid] = path_or_file_content
+        new_ref = path_or_file_content
+        if GlobalSwitchOption.BIGGER_DATA_SAVE_TO_TEMP != -1:
+            if len(path_or_file_content) >= GlobalSwitchOption.BIGGER_DATA_SAVE_TO_TEMP:
+                new_ref = join(create_safe_tempdir(), fid)
+                with open(new_ref, "wb") as f:
+                    f.write(path_or_file_content)
+            __files_dict[fid] = new_ref
+            delete_file_task(fid)
+
+        if new_ref not in list(__files_dict.values()):
+            __files_dict[fid] = new_ref
+            delete_file_task(fid)
         else:
-            return f"/file/{list(__files_dict.keys())[list(__files_dict.values()).index(path_or_file_content)]}"
+            return f"/file/{list(__files_dict.keys())[list(__files_dict.values()).index(new_ref)]}"
         return result
     if not is_valid_uri(path_or_file_content):
         fid = uuid4().hex + splitext(path_or_file_content)[1]
@@ -47,6 +81,7 @@ def get_real_uri(path_or_file_content: str | bytes) -> str:
         abs_path = abspath(path_or_file_content)
         if abs_path not in list(__files_dict.values()):
             __files_dict[fid] = abs_path
+            delete_file_task(fid)
         else:
             return f"/file/{list(__files_dict.keys())[list(__files_dict.values()).index(abs_path)]}"
         return result
